@@ -9,31 +9,26 @@
 #  https://github.com/scottzach1/python-injector-framework
 
 import functools
-from typing import Callable, TypeVar
+from collections.abc import Callable
+from typing import TypeVar
 
 from pif.providers.provider import Provider
 from pif.providers.util import intercept_args
 
-__all__ = ("Singleton",)
+__all__ = ("Factory",)
 
 T = TypeVar("T")
-UNSET = object()
 
 
-class Singleton(Provider[T]):
+class Factory(Provider):
     """
-    Provide a singleton instance.
-
-    Note that overriding any provider arguments will not cause the singleton to reevaluate.
+    Generate a new instance every call.
     """
 
-    __slots__ = ("_func", "_func", "_result", "_depends")
+    __slots__ = ("_func", "_depends")
 
     def __init__(self, func: Callable[..., T], *args, **kwargs):
         self._func = functools.partial(intercept_args(func), *args, **kwargs)
-        self._result = UNSET
 
     def _evaluate(self) -> T:
-        if self._result is UNSET:
-            self._result = self._func()
-        return self._result
+        return self._func()
